# ARROW-2828: [JS] Refactor Data, Vectors, Visitor, Typings, build, tests, dependencies