--- a/js/.gitignore
+++ b/js/.gitignore
@@ -23,7 +23,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-.vscode
+.vscode/**
+!.vscode/launch.json
 
 # Runtime data
 pids
@@ -78,10 +79,14 @@ yarn.lock
 .env
 
 # compilation targets
+doc
 dist
 targets
 
 # test data files
-test/data/
+test/data/cpp
+test/data/java
+test/data/json
+
 # jest snapshots (too big)
 test/__snapshots__/