--- a/js/gulp/arrow-task.js
+++ b/js/gulp/arrow-task.js
@@ -16,24 +16,22 @@
 // under the License.
 
 const {
-    mainExport, gCCLanguageNames,
     targetDir, observableFromStreams
 } = require('./util');
 
 const del = require('del');
 const gulp = require('gulp');
-const path = require('path');
 const { promisify } = require('util');
 const gulpRename = require(`gulp-rename`);
 const { memoizeTask } = require('./memoize-task');
 const exec = promisify(require('child_process').exec);
 const { Observable, ReplaySubject } = require('rxjs');
 
-const arrowTask = ((cache) => memoizeTask(cache, function copyMain(target, format) {
+const arrowTask = ((cache) => memoizeTask(cache, function copyMain(target) {
     const out = targetDir(target);
     const dtsGlob = `${targetDir(`es2015`, `cjs`)}/**/*.ts`;
     const cjsGlob = `${targetDir(`es2015`, `cjs`)}/**/*.js`;
-    const esmGlob = `${targetDir(`es2015`, `esm`)}/**/*.js`;
+    const esmGlob = `${targetDir(`esnext`, `esm`)}/**/*.js`;
     const es5UmdGlob = `${targetDir(`es5`, `umd`)}/*.js`;
     const es5UmdMaps = `${targetDir(`es5`, `umd`)}/*.map`;
     const es2015UmdGlob = `${targetDir(`es2015`, `umd`)}/*.js`;
@@ -46,7 +44,7 @@ const arrowTask = ((cache) => memoizeTask(cache, function copyMain(target, forma
       observableFromStreams(gulp.src(esmGlob), ch_ext(`.mjs`), gulp.dest(out)), // copy es2015 esm files and rename to `.mjs`
       observableFromStreams(gulp.src(es5UmdGlob), append(`.es5.min`), gulp.dest(out)), // copy es5 umd files and add `.min`
       observableFromStreams(gulp.src(es5UmdMaps),                     gulp.dest(out)), // copy es5 umd sourcemap files, but don't rename
-      observableFromStreams(gulp.src(es2015UmdGlob), append(`.es2015.min`), gulp.dest(out)), // copy es2015 umd files and add `.es6.min`
+      observableFromStreams(gulp.src(es2015UmdGlob), append(`.es2015.min`), gulp.dest(out)), // copy es2015 umd files and add `.es2015.min`
       observableFromStreams(gulp.src(es2015UmdMaps),                        gulp.dest(out)), // copy es2015 umd sourcemap files, but don't rename
     ).publish(new ReplaySubject()).refCount();
 }))({});