@trxcllnt 2019-01-04
@xhochy I've tried doing `git rebase apache/master`, but each time it rebases all the way back to the initial commit, then fails with a merge conflict on the `README.md` file. I'm happy to give it another shot if you have any tips (also available in slack).