@wesm 2019-01-05
I can look at fixing this tomorrow. How much do you care about maintaining the commit history?