@wesm 2019-01-05
Done. Please `git reset --hard` to that commit