@trxcllnt 2019-01-05
Thanks @wesm, I pulled and it all looks good. We can do feedback in followup PRs or discussions on the mailing list. Let me know if I can help point out where things are, or if you have any questions about what's going on.

The [`generated-data-tests`](https://github.com/trxcllnt/arrow/blob/js-data-refactor/js/test/unit/generated-data-tests.ts#L23) are a great place to start. In addition to [comprehensive tests](https://github.com/trxcllnt/arrow/blob/js-data-refactor/js/test/unit/ipc/reader/stream-reader-tests.ts#L50) across all the kinds of I/O primitives we want to support, I've tried to add as many real-world scenarios as possible (like [reading](https://github.com/trxcllnt/arrow/blob/js-data-refactor/js/test/unit/ipc/reader/streams-node-tests.ts#L116) and [writing](https://github.com/trxcllnt/arrow/blob/js-data-refactor/js/test/unit/ipc/writer/streams-node-tests.ts#L233) multiple tables to the same stream), which come from use cases at Graphistry. If you're curious about how the I/O tests work, check out [`ipc/helpers.ts`](https://github.com/trxcllnt/arrow/blob/js-data-refactor/js/test/unit/ipc/helpers.ts). That's helper functions to generate and validate each I/O primitive.

Here's a checklist to build and test locally:
* install/update to node v11.x ([`nvm install node`](https://github.com/creationix/nvm) will update to the latest)
* `$ cd arrow/js && rm -rf node_modules && npm install` to ensure libraries are up to date
* `$ npx gulp` will clean, build, and test everything

While there are always things to improve, I feel good this is ready to merge. I've been updating the Graphistry codebase to use the new APIs with no issues, and I've [started work](https://github.com/trxcllnt/fastify-arrow) ([tweet](https://twitter.com/inlineptx/status/1081535471044354048)) on libraries to integrate with other tools in the node ecosystem.