@TheNeuralBit 2019-01-07
Thanks for this Paul!
I've just been going through some smoke tests for now and everything looks good to me so far. I ran the benchmarks and noticed some changes. I posted the output from a run on master and a run on the refactor branch [here](https://gist.github.com/TheNeuralBit/9824ca1c0cfa016144fec96564e9d6ae).

Some notable changes (master -> refactor):
Accessing numeric vector by index - (49.81 ms -> 81.31 ms)
Accessing dictionary-encoded utf8 vector by index - (2758.32 ms -> 824.92 ms) :+1: 
Iterating over dictionary-encoded utf8 vector - (3152 ms -> 583.7 ms) :+1: 
slice toArray numeric vector - (1.34 ms -> 1.1 ms) :+1: 
slice toArray dictionary-encoded utf8 vector - (3147.78 ms -> 659.42 ms) :+1: 
DataFrame countBy dictionary-encoded utf8 vector - (9.95 ms -> 17.24 ms)
DataFrame filter-scan dictionary-encoded utf8 vector - (25.62ms -> 31.9ms)

Interestingly there seems to be a huge improvement in anything that decodes utf8 data. Is that because we're [using Buffer.fromString to decode utf8](https://github.com/apache/arrow/pull/3290/files#diff-8070df59d1295d866e11228d0ef6cea5R25) in node now?

There are some minor regressions elsewhere, but nothing to be too concerned about. Might be worth trying to tackle in some follow-up PRs though.