@wesm 2019-01-07
@TheNeuralBit if you could let me know when you're +1 on merging this, I can go ahead. I am swamped this week so don't want to hold this up, and I will do my best to leave some comments when I can for post-merge follow up work